Move margin code into its own namespace
