+++ src/GitHub.InlineReviews/Tags/InlineCommentTagger.cs
@@ -9,6 +9,7 @@ using GitHub.Extensions;
 using GitHub.Logging;
 using GitHub.Models;
 using GitHub.Services;
+using GitHub.InlineReviews.Margins;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Text.Tagging;
